Adding config to allow changing number of samples, seed and concurrency

+++ performUQ/DakotaFEM/DakotaFEM.py
@@ -6,17 +6,49 @@ import subprocess
 from preprocessJSON import preProcessDakota
 import platform
 import stat
+import argparse
+
+#Input Arguments Specifications
+uqArgParser = argparse.ArgumentParser()
+uqArgParser.add_argument("-filenameBIM", required=True, help="Path to BIM file")
+uqArgParser.add_argument("-filenameSAM", required=True, help="Path to SAM file")
+uqArgParser.add_argument("-filenameEVENT", required=True, help="Path to EVENT file")
+uqArgParser.add_argument("-filenameEDP", required=True, help="Path to EDP file")
+uqArgParser.add_argument("-filenameSIM", required=True, help="Path to SIM file")
+uqArgParser.add_argument("-filenameLOSS", required=True, help="Path to Damage & Loss file")
+uqArgParser.add_argument("-driverFile", required=True, help="Path to UQ driver")
+uqArgParser.add_argument("-samples", type=int, required=True, help="Number of samples")
+uqArgParser.add_argument("-seed", type=int, default=1, help="The seed used for random number generation")
+uqArgParser.add_argument("-keepSamples", help="A flag to specify whether to clean working directory for samples")
+uqArgParser.add_argument("-concurrency", type=int, default=1, help="Number of concurrent evaluations")
+
+#Parse Arguments
+uqArgs = uqArgParser.parse_args()
 
 #Reading input arguments
-bimName = sys.argv[2]
-samName = sys.argv[4]
-evtName = sys.argv[6]
-edpName = sys.argv[8]
-lossName = sys.argv[10]
-simName = sys.argv[12]
-driverFile = sys.argv[14]
+bimName = uqArgs.filenameBIM
+samName = uqArgs.filenameSAM
+evtName = uqArgs.filenameEVENT
+edpName = uqArgs.filenameEDP
+lossName = uqArgs.filenameLOSS
+simName = uqArgs.filenameSIM
+driverFile = uqArgs.driverFile
 
-numSamples = 5 #TODO: needs to be configured
+concurrency = 1
+if uqArgs.concurrency is not None:
+    concurrency = uqArgs.concurrency
+
+doCleanup = True
+if uqArgs.keepSamples.lower() == "true":
+    doCleanup = False
+elif  uqArgs.keepSamples.lower() == "false" or uqArgs.keepSamples is None:
+    doCleanup = True
+else:
+    doCleanup = True
+    print "Warning: {} is not a valid value for keepSamples".format(uqArgs.keepSamples)
+
+numSamples = uqArgs.samples
+rngSeed = uqArgs.seed
 
 bldgName = bimName[:-9]
 
@@ -33,7 +65,7 @@ os.mkdir(bldgName)
 
 #Run Preprocess for Dakota
 scriptDir = os.path.dirname(os.path.realpath(__file__))
-numRVs = preProcessDakota(bimName, evtName, samName, edpName, lossName, simName, driverFile, bldgName, numSamples)
+numRVs = preProcessDakota(bimName, evtName, samName, edpName, lossName, simName, driverFile, bldgName, numSamples, rngSeed, concurrency)
 
 
 #Create Template Directory and copy files
@@ -64,4 +96,5 @@ else:
 subprocess.Popen(postprocessCommand, shell=True).wait()
 
 #Clean up building folder
-shutil.rmtree(bldgName)
+if doCleanup:
+    shutil.rmtree(bldgName)
